Splits up clients into separate features for wasm-compatibility
