+++ fedimint-bitcoind/Cargo.toml
@@ -16,9 +16,9 @@ anyhow = "1.0.66"
 async-trait = "*"
 bitcoin = "0.29.2"
 bitcoin_hashes = "0.11.0"
-bitcoincore-rpc = "0.16.0"
-electrum-client = "0.12.0"
-esplora-client = { version = "0.5.0", default-features = false, features = ["async", "async-https-rustls"] }
+bitcoincore-rpc = { version = "0.16.0", optional = true }
+electrum-client = {version = "0.12.0", optional = true }
+esplora-client = { version = "0.5.0", default-features = false, features = ["async", "async-https-rustls"], optional = true }
 lazy_static = "1.4.0"
 fedimint-core  = { path = "../fedimint-core" }
 fedimint-logging = { path = "../fedimint-logging" }
@@ -26,3 +26,6 @@ rand = "0.8"
 serde = { version = "1.0.149", features = [ "derive" ] }
 tracing = "0.1.37"
 url = "2.3.1"
+
+[features]
+default = ["bitcoincore-rpc", "electrum-client", "esplora-client"]
